MDWFP-73 addressing review comments.
